inclusion of graphics and API
