+++ api/urls.py
@@ -7,6 +7,7 @@ from api.api_admin import AdminViewset
 from api.students.angola_bernardo import AngolaBernardoViewset
 from api.students.angola_estima import AngolaEstimaViewset
 from api.students.angola_lupossa import AngolaLupossaViewset
+from api.students.mocambique_momade import MocambiqueMomadeViewset
 
 router = DefaultRouter()
 
@@ -17,8 +18,12 @@ router.register(r'budgets', BudgetViewset, basename='budgets')
 router.register(r'angola-estima', AngolaEstimaViewset, basename='angola_estima_api')
 router.register(r'angola-lupossa', AngolaLupossaViewset, basename='angola_lupossa_api')
 router.register(r'angola-bernardo', AngolaBernardoViewset, basename='angola_bernardo_api')
+router.register(r'mocambique-momade', MocambiqueMomadeViewset, basename='mocambique_momade_api')
+
 
 urlpatterns = [
     url(r'^', include(router.urls)),
     url(r'^api-auth/', include('rest_framework.urls'))
 ]
+
+
