+++ config/settings.py
@@ -130,7 +130,7 @@ LANGUAGE_CODE = 'pt'
 TIME_ZONE = 'UTC'
 USE_I18N = True
 USE_L10N = True
-USE_TZ = True
+# USE_TZ = True
 LANGUAGES = [
     ('pt', _('Portuguese')),
     ('en', _('English')),
