+++ frontend/student_views/mocambique_momade.py
@@ -1,0 +1,14 @@
+from django.shortcuts import render
+from django.views import View
+
+from frontend.forms import BudgetPerCountryForm
+
+
+class YearPercentGetParameter(View):
+    def get(self, request):
+        form = BudgetPerCountryForm()
+        ctx = {
+            'form': form
+        }
+
+        return render(request, 'frontend/students/moz_momade.html', context=ctx)
