+++ frontend/templates/frontend/students/moz_momade.html
@@ -1,0 +1,35 @@
+{% extends 'frontend/base.html' %}
+
+{%load static %}
+
+{% block extrahead %}
+    <script src="https://cdn.amcharts.com/lib/4/core.js"></script>
+    <script src="https://cdn.amcharts.com/lib/4/charts.js"></script>
+    <script src="https://cdn.amcharts.com/lib/4/themes/animated.js"></script>
+    <script src="{% static 'students/moz_momade.js' %}"> </script>
+{% endblock %}
+
+{% block extrastyle %}
+    <style>
+        #chartdiv {
+            width: 100%;
+            height: 500px;
+        }
+    </style>
+
+{% endblock %}
+
+{% block content %}
+
+    <center>
+        <div style="color:#4242f5;font-size:40px;">
+            Gastos Orçamentais de <strong>Moçambique <i>por ano</i> económico </strong>! <br>
+        </div>
+
+    </center>
+
+
+
+    <div id="chartdiv"></div>
+
+{% endblock %}
