+++ requirements.txt
@@ -16,8 +16,8 @@ django-test-without-migrations==0.6
 djangorestframework==3.11.0
 djangorestframework-recursive==0.1.2
 gunicorn==20.0.4
-Pillow==8.3.0
-psycopg2-binary==2.8.6
+Pillow==9.1.0
+psycopg2-binary==2.9.3
 py-moneyed==0.8.0
 requests==2.23.0
 sentry-sdk==0.19.2
